Derive Init from Serialize, Deserialize
As preparation for the refactoring with a initial fork prior to the
runtime setup, the Init structure must be sendalbe via a pipe. This is
best accomplished by deriving it from serde's Serialize and Deserialize.
