+++ northstar/src/runtime/process/mod.rs
@@ -9,7 +9,7 @@ use super::{
 };
 use crate::{
     common::{container::Container, non_null_string::NonNullString},
-    npk::manifest::{Capability, Manifest, RLimitResource, RLimitValue},
+    npk::manifest::Manifest,
     runtime::{
         console::{self, Peer},
         error::Context,
@@ -17,7 +17,6 @@ use crate::{
     seccomp,
 };
 use async_trait::async_trait;
-use caps::CapsHashSet;
 use futures::{future::ready, Future, FutureExt};
 use log::{debug, error, info, warn};
 use nix::{
@@ -26,7 +25,7 @@ use nix::{
     unistd,
 };
 use std::{
-    collections::{HashMap, HashSet},
+    collections::HashMap,
     ffi::{c_void, CString},
     fmt,
     mem::forget,
@@ -136,13 +135,13 @@ impl Launcher {
         )
         .await;
 
-        let capabilities = capabilities(manifest);
+        let capabilities = manifest.capabilities.clone();
         let fds = fds.drain().collect::<Vec<_>>();
         let uid = manifest.uid;
         let gid = manifest.gid;
         let groups = groups(manifest);
         let mounts = fs::prepare_mounts(&self.config, root, manifest).await?;
-        let rlimits = rlimits(manifest);
+        let rlimits = manifest.rlimits.clone();
         let root = root.to_owned();
         let seccomp = seccomp_filter(manifest);
 
@@ -513,113 +512,6 @@ fn seccomp_filter(manifest: &Manifest) -> Option<seccomp::AllowList> {
         ));
     }
     None
-}
-
-/// Capability settings applied in init
-struct Capabilities {
-    all: CapsHashSet,
-    bounded: CapsHashSet,
-    set: CapsHashSet,
-}
-
-/// Calculate capability sets
-fn capabilities(manifest: &Manifest) -> Capabilities {
-    let all = caps::all();
-    let mut bounded =
-        caps::read(None, caps::CapSet::Bounding).expect("Failed to read bounding caps");
-    // Convert the set from the manifest to a set of caps::Capbility
-    let set = manifest
-        .capabilities
-        .clone()
-        .unwrap_or_default()
-        .into_iter()
-        .map(Into::into)
-        .collect::<HashSet<caps::Capability>>();
-    bounded.retain(|c| !set.contains(c));
-    Capabilities { all, bounded, set }
-}
-
-impl From<Capability> for caps::Capability {
-    fn from(cap: Capability) -> Self {
-        match cap {
-            Capability::CAP_CHOWN => caps::Capability::CAP_CHOWN,
-            Capability::CAP_DAC_OVERRIDE => caps::Capability::CAP_DAC_OVERRIDE,
-            Capability::CAP_DAC_READ_SEARCH => caps::Capability::CAP_DAC_READ_SEARCH,
-            Capability::CAP_FOWNER => caps::Capability::CAP_FOWNER,
-            Capability::CAP_FSETID => caps::Capability::CAP_FSETID,
-            Capability::CAP_KILL => caps::Capability::CAP_KILL,
-            Capability::CAP_SETGID => caps::Capability::CAP_SETGID,
-            Capability::CAP_SETUID => caps::Capability::CAP_SETUID,
-            Capability::CAP_SETPCAP => caps::Capability::CAP_SETPCAP,
-            Capability::CAP_LINUX_IMMUTABLE => caps::Capability::CAP_LINUX_IMMUTABLE,
-            Capability::CAP_NET_BIND_SERVICE => caps::Capability::CAP_NET_BIND_SERVICE,
-            Capability::CAP_NET_BROADCAST => caps::Capability::CAP_NET_BROADCAST,
-            Capability::CAP_NET_ADMIN => caps::Capability::CAP_NET_ADMIN,
-            Capability::CAP_NET_RAW => caps::Capability::CAP_NET_RAW,
-            Capability::CAP_IPC_LOCK => caps::Capability::CAP_IPC_LOCK,
-            Capability::CAP_IPC_OWNER => caps::Capability::CAP_IPC_OWNER,
-            Capability::CAP_SYS_MODULE => caps::Capability::CAP_SYS_MODULE,
-            Capability::CAP_SYS_RAWIO => caps::Capability::CAP_SYS_RAWIO,
-            Capability::CAP_SYS_CHROOT => caps::Capability::CAP_SYS_CHROOT,
-            Capability::CAP_SYS_PTRACE => caps::Capability::CAP_SYS_PTRACE,
-            Capability::CAP_SYS_PACCT => caps::Capability::CAP_SYS_PACCT,
-            Capability::CAP_SYS_ADMIN => caps::Capability::CAP_SYS_ADMIN,
-            Capability::CAP_SYS_BOOT => caps::Capability::CAP_SYS_BOOT,
-            Capability::CAP_SYS_NICE => caps::Capability::CAP_SYS_NICE,
-            Capability::CAP_SYS_RESOURCE => caps::Capability::CAP_SYS_RESOURCE,
-            Capability::CAP_SYS_TIME => caps::Capability::CAP_SYS_TIME,
-            Capability::CAP_SYS_TTY_CONFIG => caps::Capability::CAP_SYS_TTY_CONFIG,
-            Capability::CAP_MKNOD => caps::Capability::CAP_MKNOD,
-            Capability::CAP_LEASE => caps::Capability::CAP_LEASE,
-            Capability::CAP_AUDIT_WRITE => caps::Capability::CAP_AUDIT_WRITE,
-            Capability::CAP_AUDIT_CONTROL => caps::Capability::CAP_AUDIT_CONTROL,
-            Capability::CAP_SETFCAP => caps::Capability::CAP_SETFCAP,
-            Capability::CAP_MAC_OVERRIDE => caps::Capability::CAP_MAC_OVERRIDE,
-            Capability::CAP_MAC_ADMIN => caps::Capability::CAP_MAC_ADMIN,
-            Capability::CAP_SYSLOG => caps::Capability::CAP_SYSLOG,
-            Capability::CAP_WAKE_ALARM => caps::Capability::CAP_WAKE_ALARM,
-            Capability::CAP_BLOCK_SUSPEND => caps::Capability::CAP_BLOCK_SUSPEND,
-            Capability::CAP_AUDIT_READ => caps::Capability::CAP_AUDIT_READ,
-            Capability::CAP_PERFMON => caps::Capability::CAP_PERFMON,
-            Capability::CAP_BPF => caps::Capability::CAP_BPF,
-            Capability::CAP_CHECKPOINT_RESTORE => caps::Capability::CAP_CHECKPOINT_RESTORE,
-        }
-    }
-}
-
-type RLimits = HashMap<rlimit::Resource, RLimitValue>;
-
-fn rlimits(manifest: &Manifest) -> RLimits {
-    manifest
-        .rlimits
-        .as_ref()
-        .map(|l| {
-            l.iter()
-                .map(|(k, v)| {
-                    let resource = match k {
-                        RLimitResource::AS => rlimit::Resource::AS,
-                        RLimitResource::CORE => rlimit::Resource::CORE,
-                        RLimitResource::CPU => rlimit::Resource::CPU,
-                        RLimitResource::DATA => rlimit::Resource::DATA,
-                        RLimitResource::FSIZE => rlimit::Resource::FSIZE,
-                        RLimitResource::LOCKS => rlimit::Resource::LOCKS,
-                        RLimitResource::MEMLOCK => rlimit::Resource::MEMLOCK,
-                        RLimitResource::MSGQUEUE => rlimit::Resource::MSGQUEUE,
-                        RLimitResource::NICE => rlimit::Resource::NICE,
-                        RLimitResource::NOFILE => rlimit::Resource::NOFILE,
-                        RLimitResource::NPROC => rlimit::Resource::NPROC,
-                        RLimitResource::RSS => rlimit::Resource::RSS,
-                        RLimitResource::RTPRIO => rlimit::Resource::RTPRIO,
-                        #[cfg(not(target_os = "android"))]
-                        RLimitResource::RTTIME => rlimit::Resource::RTTIME,
-                        RLimitResource::SIGPENDING => rlimit::Resource::SIGPENDING,
-                        RLimitResource::STACK => rlimit::Resource::STACK,
-                    };
-                    (resource, v.clone())
-                })
-                .collect()
-        })
-        .unwrap_or_default()
 }
 
 // Set the child subreaper flag of the calling thread
